+++ InOut/libmpadec/mp3dec.c
@@ -104,7 +104,7 @@ int32_t mp3dec_init_file(mp3dec_t mp3dec, FILE *f, int64_t length, int32_t nogap
       }
       (void) fseek(f, mp3->stream_offset, SEEK_SET);
     } else mp3->flags &= ~MP3DEC_FLAG_SEEKABLE; 
-    r = fread(mp3->in_buffer, 1, 4, f);
+    r = (int32_t) fread(mp3->in_buffer, 1, 4, f);
     if (r < 4) {
       mp3dec_reset(mp3);
       return ((r < 0) ? MP3DEC_RETCODE_INVALID_PARAMETERS :
@@ -117,7 +117,7 @@ int32_t mp3dec_init_file(mp3dec_t mp3dec, FILE *f, int64_t length, int32_t nogap
       mp3->flags &= ~MP3DEC_FLAG_SEEKABLE;
       if (mp3->stream_size && (n > (mp3->stream_size - mp3->in_buffer_used)))
         n = (int32_t)(mp3->stream_size - mp3->in_buffer_used);
-      n = fread(mp3->in_buffer + mp3->in_buffer_used, 1, n, f);
+      n =  (int32_t) fread(mp3->in_buffer + mp3->in_buffer_used, 1, n, f);
       if (n < 0) n = 0;
       mp3->in_buffer_used += n;
       mp3->stream_position = mp3->in_buffer_used;
@@ -126,7 +126,7 @@ int32_t mp3dec_init_file(mp3dec_t mp3dec, FILE *f, int64_t length, int32_t nogap
       int32_t n = sizeof(mp3->in_buffer);
       if (mp3->stream_size && (n > mp3->stream_size))
         n = (int32_t)mp3->stream_size;
-      n = fread(mp3->in_buffer, 1, n, f);
+      n =  (int32_t) fread(mp3->in_buffer, 1, n, f);
       if (n < 0) n = 0;
       mp3->stream_position = mp3->in_buffer_used = n;
     }
@@ -156,7 +156,7 @@ int32_t mp3dec_init_file(mp3dec_t mp3dec, FILE *f, int64_t length, int32_t nogap
           int32_t n = sizeof(mp3->in_buffer);
           if (mp3->stream_size && (n > mp3->stream_size))
             n = (int32_t)mp3->stream_size;
-          n = fread(mp3->in_buffer, 1, n, f);
+          n =  (int32_t) fread(mp3->in_buffer, 1, n, f);
           if (n <= 0){ /* n = 0; */ break; } /* EOF */
           mp3->stream_position = mp3->in_buffer_used = n;
           r = mpadec_decode(mp3->mpadec, mp3->in_buffer,
@@ -316,7 +316,7 @@ int32_t mp3dec_decode(mp3dec_t mp3dec, uint8_t *buf, uint32_t bufsize, uint32_t 
       n = sizeof(mp3->in_buffer) - mp3->in_buffer_used;
       if (mp3->stream_size && (n > (mp3->stream_size - mp3->stream_position)))
         n = (int32_t)(mp3->stream_size - mp3->stream_position);
-      if (n) r = fread(mp3->in_buffer + mp3->in_buffer_used, 1, n, mp3->f);
+      if (n) r =  (int32_t) fread(mp3->in_buffer + mp3->in_buffer_used, 1, n, mp3->f);
       else r = 0;
       if (r < 0) r = 0;
       mp3->in_buffer_used += r;
