+++ OOps/ugrw1.c
@@ -113,7 +113,7 @@ int32_t elapsedtime(CSOUND *csound, RDTIME *p)
 int32_t instimset(CSOUND *csound, RDTIME *p)
 {
    IGN(csound);
-    p->instartk = CS_KCNT;
+   p->instartk = CS_KCNT;
     *p->rslt = FL(0.0);
     return OK;
 }
@@ -756,7 +756,7 @@ int32_t printk3(CSOUND *csound, PRINTK3 *p)
 }
 
 #include "../Opcodes/zak.h"
-static int32_t GetZaBounds(CSOUND *csound, MYFLT **zastart){
+static int64_t GetZaBounds(CSOUND *csound, MYFLT **zastart){
     ZAK_GLOBALS *zz;
     zz = (ZAK_GLOBALS*) csound->QueryGlobalVariable(csound, "_zak_globals");
     if (zz==NULL) {
@@ -777,7 +777,7 @@ int32_t inz(CSOUND *csound, IOZ *p)
     uint32_t n, nsmps = CS_KSMPS;
     /* Check to see this index is within the limits of za space.     */
     MYFLT* zastart;
-    int32_t zalast = GetZaBounds(csound, &zastart);
+    int64_t zalast = GetZaBounds(csound, &zastart);
     indx = (int32_t) *p->ndx;
     if (UNLIKELY(indx + nchns >= zalast)) goto err1;
     else if (UNLIKELY(indx < 0)) goto err2;
@@ -813,7 +813,7 @@ int32_t outz(CSOUND *csound, IOZ *p)
 
     /* Check to see this index is within the limits of za space.    */
     MYFLT* zastart;
-    int32_t zalast = GetZaBounds(csound, &zastart);
+    int64_t zalast = GetZaBounds(csound, &zastart);
     indx = (int32) *p->ndx;
     if (UNLIKELY((indx + nchns) >= zalast)) goto err1;
     else if (UNLIKELY(indx < 0)) goto err2;
