+++ include/csoundCore.h
@@ -1660,11 +1660,11 @@ static inline double PHMOD1(double p) {
                           int32_t fd, int32_t mode, SFLIB_INFO *sfinfo,
                           int32_t close_desc);
     int32_t (*SndfileClose)(CSOUND *csound, void *);
-    int32_t (*SndfileWrite)(CSOUND *, void *, MYFLT *, int32_t);
-    int32_t (*SndfileRead)(CSOUND *, void *, MYFLT *, int32_t);
+    int64_t (*SndfileWrite)(CSOUND *, void *, MYFLT *, int64_t);
+    int64_t (*SndfileRead)(CSOUND *, void *, MYFLT *, int64_t);
     int64_t (*SndfileWriteSamples)(CSOUND *, void *, MYFLT *, int64_t);
     int64_t (*SndfileReadSamples)(CSOUND *, void *, MYFLT *, int64_t);
-    int32_t (*SndfileSeek)(CSOUND *, void *, int32_t, int32_t);
+    int64_t (*SndfileSeek)(CSOUND *, void *, int64_t, int32_t);
     int32_t (*SndfileSetString)(CSOUND *csound, void *sndfile, int32_t str_type, const char* str);
     const char *(*SndfileStrError)(CSOUND *csound, void *);
     int32_t (*SndfileCommand)(CSOUND *, void *, int32_t , void *, int32_t );
